FLask-Sql-server/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
langchain-google-genai
flask-cors

python-dotenv